+++ examples/djangoapp/chat/views.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.shortcuts import render_to_response
 from django.template import RequestContext