+++ pulsar/apps/pulse/__init__.py
@@ -1,0 +1,15 @@
+'''The :mod:`pulsar.apps.pulse` module is a django application for running
+a django web site with pulsar. Add it to the list of your ``INSTALLED_APPS``::
+
+    INSTALLED_APPS = (
+        ...,
+        'pulsar.apps.pulse',
+        ...
+    )
+    
+and run the site via the ``pulse`` command::
+
+    python manage.py pulse
+
+
+'''