+++ pulsar/apps/wsgi/handlers.py
@@ -98,7 +98,6 @@ from pulsar import Http404, PermissionDenied, HttpException, async
 from .route import Route
 from .utils import wsgi_request
 from .content import Html
-from .wrappers import WsgiResponseGenerator
 
 __all__ = ['WsgiHandler', 'LazyWsgi', 'Router',
            'MediaRouter', 'FileRouter', 'MediaMixin']
@@ -266,7 +265,7 @@ request, the ``get(self, request)`` method must be implemented.
             if callable is None:
                 raise HttpException(status=405,
                                     msg='Method "%s" not allowed' % method)
-            return WsgiResponseGenerator(request, callable)
+            return callable(request)
         
     def resolve(self, path, urlargs=None):
         '''Resolve a path and return a ``(handler, urlargs)`` tuple or